refactor(types): type repository Object params + Pref getters (SPES-014)
T2: fix .obs dynamic dispatch in pl_player superResolutionType ternary
T3: type 31 video/live repository params as String/int (full chain sync)
T4: type 18 dynamics repository params as String/int (full chain sync)
T5: type 34 Pref dynamic getters; move B站 enum getters to BiliPref
    (removes adapter imports from lib/utils); drop redundant casts/imports

+++ lib/adapters/bilibili/pages/audio/controller.dart
@@ -87,7 +87,7 @@ class AudioController extends GetxController
 
   late double speed = 1.0;
 
-  late final Rx<PlayRepeat> playMode = PlayRepeat.values[Pref.audioPlayMode as int].obs;
+  late final Rx<PlayRepeat> playMode = PlayRepeat.values[Pref.audioPlayMode].obs;
 
   @override
   late final isLogin = Accounts.main.isLogin;
