+++ lib/adapters/bilibili/pages/setting/pages/color_select.dart
@@ -219,5 +219,5 @@ class _ColorSelectPageState extends State<ColorSelectPage> {
 class _ColorSelectController extends GetxController {
   final RxBool dynamicColor = (Pref.dynamicColor).obs;
   final RxInt currentColor = (Pref.customColor).obs;
-  final Rx<ThemeType> themeType = ThemeType.values[Pref.themeType as int].obs;
+  final Rx<ThemeType> themeType = ThemeType.values[Pref.themeType].obs;
 }
